# Use Seek.stream_position instead of seek(0).