@zamazan4ik 2022-04-11
I am a little wondered why in previous versions, `seek(0)` was used. Seems really strange to me