@PJB3005 2022-04-11
`stream_position()` wasn't added until Rust 1.51, to my massive surprise. So that's probably why.