@PJB3005 2022-04-11
I don't have the most scientific benchmark setup here because I just ran some ETW profiles when I noticed my zip files (like ~50 MB, ~13k mostly-tiny files) took way longer to open than they should on debug mode, then cross-referenced the code with the amount of seek calls and yep.

I just made a quick and dirty Read + Seek impl that doesn't pass through `io::SeekFrom::Current(0)` to the base stream:
```rust
struct FTellStream<R> {
    inner: R,
}

impl<R: Read + Seek> FTellStream<R> {
    pub fn new(inner: R) -> Self {
        FTellStream { inner }
    }
}

impl<R: Read + Seek> Read for FTellStream<R> {
    fn read(&mut self, buf: &mut [u8]) -> Result<usize, io::Error> {
        self.inner.read(buf)
    }

    fn read_exact(&mut self, buf: &mut [u8]) -> io::Result<()> {
        self.inner.read_exact(buf)
    }
}

impl<R: Read + Seek> Seek for FTellStream<R> {
    fn seek(&mut self, from: io::SeekFrom) -> Result<u64, io::Error> {
        if from == io::SeekFrom::Current(0) {
            return self.inner.stream_position();
        }

        self.inner.seek(from)
    }
}
```

Without, doing this on these zip files:
```rust
        let a = Instant::now();
        let file = fs::File::open(zip_path).unwrap();
        let mut zip = ZipArchive::new( BufReader::new(file)).unwrap();
        let b = Instant::now();

        println!("{}, {:?}", zip.len(), b.duration_since(a));
```

I get like:

```
17092, 85.636ms
17092, 86.8534ms
17092, 84.1522ms
17076, 82.9949ms
17092, 84.8181ms
17092, 84.4906ms
```

With:

```rust
        let a = Instant::now();
        let file = fs::File::open(zip_path).unwrap();
        let mut zip = ZipArchive::new(FTellStream::new(BufReader::new(file))).unwrap();
        let b = Instant::now();

        println!("{}, {:?}", zip.len(), b.duration_since(a));
```

```
17092, 38.4624ms
17092, 38.2349ms
17092, 40.5351ms
17076, 38.3368ms
17092, 37.9675ms
17092, 37.6783ms
```

Should be noted that this is on Windows, for however much that matters.